+++ src/components/index.ts
@@ -4,7 +4,7 @@ export { default as PageIntro } from './PageIntro'
 export { default as Profile } from './Profile'
 export { default as UserList } from './UserList'
 export { default as PersonalAnswer } from './PersonalAnswer'
-export { default as PassWordInput } from './PasswordInput'
+export { default as PassWordInput } from './PassWordInput'
 export { default as IconButton } from './IconButton'
 export { default as SearchBar } from './SearchBar'
 export * from './StarRating'
